test multiple palettes output
